40-bit secure vole in Z2^32

+++ backend/src/quicksilver/vole.rs
@@ -5,18 +5,18 @@ pub struct Segments {
 }
 
 pub struct CorrSender {
-    pub xs_in: Vec<u64>,
-    pub mc_in: Vec<u64>,
-    pub xs_mul: Vec<u64>,
-    pub mc_mul: Vec<u64>,
-    pub xs_mul_check: Vec<u64>,
-    pub mc_mul_check: Vec<u64>,
+    pub xs_in: Vec<u128>,
+    pub mc_in: Vec<u128>,
+    pub xs_mul: Vec<u128>,
+    pub mc_mul: Vec<u128>,
+    pub xs_mul_check: Vec<u128>,
+    pub mc_mul_check: Vec<u128>,
 }
 
 pub struct CorrReceiver {
-    pub ks_in: Vec<u64>,
-    pub ks_mul: Vec<u64>,
-    pub ks_mul_check: Vec<u64>,
+    pub ks_in: Vec<u128>,
+    pub ks_mul: Vec<u128>,
+    pub ks_mul_check: Vec<u128>,
 }
 
 impl Segments {
